Clean up location of animation files.
